Added a linter and checks (#89)
This adds [stylelint](https://stylelint.io/) (a CSS linter) and a few checks in gulp as well as some minor changes.

__stylelint:__

Stylelint is the CSS linter I choose. It does what I want a CSS linter to do and that's it. If you know a better linter, we can switch anytime.

I configured the linter such that there are as few modifications as possible. The only things that changed in the CSS files are some whitespaces and newlines. All non-whitespace changes in the themes are errors that the linter pointed out.

To make the linter stricter, we just have re-enable the rules I disabled (all rules which are set to `null`). [Here](https://stylelint.io/user-guide/rules) is the documentation of all rules.

__gulp checks:__

I also added two checks, which 1) check that every theme has a screenshot and 2) that every theme (plus screenshot) is in the readme.

Both checks and the linter will be run by Travis. 
(Yay, Travis is actually doing something now!)

__Minor changes:__

* Fixed a typo in `package.json`.
* `npm ci`, as described in the README, doesn't work without `package-lock.json`, so I changed it to `npm i` and added the `package-lock.json` to the `.gitignore`.